% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.r
\name{check_inputs}
\alias{check_inputs}
\title{Check the validity of remote sensing classification data}
\usage{
check_inputs(map, val, reproj = FALSE)
}
\arguments{
\item{map}{classification results as a *raster* or *SpatialPolygons* object.}

\item{val}{reference data for validating the classification. Accepts *raster*, *SpatialPolygons* or *SpatialPoints* objects.}

\item{reproj}{logical flag indicating if classification data should be reprojected to match reference data.}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
This function takes classification and validation inputs, given as *raster* or *Spatial* objects, and tests if projection are the same.
}
