% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_summs.R
\name{conf_summs}
\alias{conf_summs}
\title{Calculates summary statistics and totals from a confusion matrix}
\usage{
conf_summs(confmat)
}
\arguments{
\item{confmat}{a confusion matrix created by conf_mat(), or matrix with named columns and rows.}
}
\value{
Several useful statistics.
}
\description{
This function calculates diagonal sums, row sums, and similar summaries from the confusion matrix, which are needed to calculate accuracy metrics.
}
