% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kia.R
\name{kia}
\alias{kia}
\title{Per class accuracies and *kappa* index of agreement}
\usage{
kia(confmat)
}
\arguments{
\item{confmat}{a confusion matrix created by conf_mat(), or matrix with named columns and rows.}
}
\value{
Overall Accuracy, Kappa Index of Agreement, Per-class Omission and Comission Errors
}
\description{
Calculate per-class Omission and Comission errors and the Kappa index of agreement.
}
