% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pontius.R
\name{pontius}
\alias{pontius}
\title{Calculates quantity and allocation disagreement indexes.}
\usage{
pontius(confmat)
}
\arguments{
\item{confmat}{a confusion matrix created by conf_mat(), or matrix with named columns and rows.}
}
\value{
Overall agreement and disagreement, Allocation Disagreement, Quantity Disagreement.
}
\description{
function to calculate the disagreement measures from:
Pontius Jr, Robert Gilmore, and Marco Millones. "Death to Kappa: birth of quantity disagreement and allocation disagreement for accuracy assessment." International Journal of Remote Sensing 32.15 (2011): 4407-4429.
}
